Major work on API detection. Pattern files are converted to a source-code file much better, detection takes the current library versions into account, and patching has started!
Also added a 'release' version of our DLL, which differs from the 'debug' version in only one thing : the DXBX_DEBUG define is off (which leaves lots of unneeded stuff out). Note, that these two DLL projects should be kept in sync with each other from now on!

As an expiriment, the 'release' DLL uses an environment variable for the search path, so we won't have to include all third party units into our projects; To use this, go to Delphi, Tools > Options > Environment Variables, and add the DXBXLIB define, let it point to your absolute path of the Dxbx\Libraries folder.

As a sidenote : The Xbox layout of the structure that the FS-register points to, is different from the win32 version. Robert will patch all FS-referencing opcodes, so they become win32-friendly. This will enable us to print stack-traces of all detected functions (most code for this is already present; I used the extension mechanism of JclDebug for this - see uDxbxDebugUtils.pas)

git-svn-id: https://svn.code.sf.net/p/dxbx/svn@260 268f5bea-7c72-4cc5-bae1-aebef56d899c