found more sound methods that are not patched.
I added these methods so we know that we still need to implement these. params and all other things are not checked.
its some sort of todo list so we dont forget them to implement.

Finding missing patches by the patternfiles.. is not fun, but its the best way of finding these missing patches.

git-svn-id: https://svn.code.sf.net/p/dxbx/svn@939 268f5bea-7c72-4cc5-bae1-aebef56d899c